move types to separate file due to failing lint
